Revert revision attribute type to i64

+++ yt/server/cypress_server/node_proxy_detail.cpp
@@ -660,7 +660,8 @@ bool TNontemplateCypressNodeProxyBase::GetBuiltinAttribute(
 
         case EInternedAttributeKey::Revision:
             BuildYsonFluently(consumer)
-                .Value(node->GetRevision());
+                // TODO(babenko): KWYT-630
+                .Value(static_cast<i64>(node->GetRevision()));
             return true;
 
         case EInternedAttributeKey::AttributesRevision:
